refactor: move swap-form thunks
